Revert "v0.0.11"
This reverts commit 1eb5d496463d5d8f4d12271e673dd57babc51d56, reversing
changes made to c22b49121d3c72454a8c358c6e27978a69472413.
