--- src/gui/__init__.py
@@ -1,20 +1,0 @@
-"""GUI components for GM2Godot."""
-from .gui import (
-    ModernButton,
-    ModernProgressBar,
-    Icon,
-    SetupUI,
-    AboutDialog,
-    get_version,
-    VERSION
-)
-
-__all__ = [
-    'ModernButton',
-    'ModernProgressBar',
-    'Icon',
-    'SetupUI',
-    'AboutDialog',
-    'get_version',
-    'VERSION'
-]
